Time: 0 ms (100%), Space: 42.3 MB (62.34%) - LeetHub

+++ 3622-check-divisibility-by-digit-sum-and-product/3622-check-divisibility-by-digit-sum-and-product.java
@@ -1,7 +1,7 @@
 class Solution {
     public boolean checkDivisibility(int n) {
-        long sum = 0 ;
-        long product = 1;
+        int sum = 0 ;
+        int product = 1;
         int chk = n;
         while(n>0)
         {
